requirements.txt
feedparser
yt-dlp
PyYAML